requirements.txt
telethon
python-decouple
aioredis[hiredis]